// wordle-frontend/src/app/api/models/sp-stats-get-advanced-result.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SpStatsGetAdvancedResult {
  avgGuessesToWin: number | null;
  globalGuess1: number;
  globalGuess2: number;
  globalGuess3: number;
  globalGuess4: number;
  globalGuess5: number;
  globalGuess6: number;
  globalWinRate: number | null;
  hardWord1: string | null;
  hardWord2: string | null;
  hardWord3: string | null;
  hardWord4: string | null;
  hardWord5: string | null;
  hardWordGames1: number;
  hardWordGames2: number;
  hardWordGames3: number;
  hardWordGames4: number;
  hardWordGames5: number;
  hardWordWinRate1: number;
  hardWordWinRate2: number;
  hardWordWinRate3: number;
  hardWordWinRate4: number;
  hardWordWinRate5: number;
  topWord1: string | null;
  topWord10: string | null;
  topWord2: string | null;
  topWord3: string | null;
  topWord4: string | null;
  topWord5: string | null;
  topWord6: string | null;
  topWord7: string | null;
  topWord8: string | null;
  topWord9: string | null;
  topWordCount1: number;
  topWordCount10: number;
  topWordCount2: number;
  topWordCount3: number;
  topWordCount4: number;
  topWordCount5: number;
  topWordCount6: number;
  topWordCount7: number;
  topWordCount8: number;
  topWordCount9: number;
  totalGamesPlayed: number | null;
  totalGamesWon: number | null;
}